requirements.txt
# tensorflow>=2.13.0,<3.0.0
# keras>=2.13.0
# numpy>=1.24.0,<2.0.0
# pandas>=2.0.0
# matplotlib>=3.7.0
# seaborn>=0.12.0
# scikit-learn>=1.3.0
# Pillow>=10.0.0
# Flask>=3.0.0
# opencv-python>=4.8.0
# tqdm>=4.66.0
# jupyter>=1.0.0


flask==3.0.0
tensorflow-cpu==2.17.0
numpy==1.26.4
pillow==10.0.0
werkzeug==3.0.0
gunicorn==21.2.0
scikit-learn==1.5.0
h5py==3.11.0
gdown
